중간 커밋

+++ java_work/ex01/src/chap03/OperatorEx6.java
@@ -3,7 +3,7 @@ class OperatorEx6 {
       public static void main(String[] args) { 
             byte a = 10; 
             byte b = 20; 
-            byte c = a + b; 
+            byte c = (byte)(a + b); 
 
             System.out.println(c); 
       }
